Organize API code
New structure:

lib/
  api/
    endpoints/
    helpers/
  record_parser/
    sorting/

spec/
  api/
    endpoints/
    helpers/
  record_parser/
    sorting/
