+++ lib/api/endpoints.rb
@@ -1,0 +1,3 @@
+require 'api/endpoints/get_by_birth_date'
+require 'api/endpoints/get_by_gender_and_last_name'
+require 'api/endpoints/get_by_last_name_descending'
