+++ lib/api/endpoints/get_by_birth_date.rb
@@ -1,0 +1,14 @@
+require 'api/helpers'
+
+module API
+  module Endpoints
+    class GetByBirthDate < Grape::API
+      helpers Helpers::SortHelper
+      helpers Helpers::StoreHelper
+
+      get :birthdate do
+        sort(records, :ByBirthDate)
+      end
+    end
+  end
+end
