+++ lib/api/endpoints/get_by_gender_and_last_name.rb
@@ -1,0 +1,14 @@
+require 'api/helpers'
+
+module API
+  module Endpoints
+    class GetByGenderAndLastName < Grape::API
+      helpers Helpers::SortHelper
+      helpers Helpers::StoreHelper
+
+      get :gender do
+        sort(records, :ByGenderAndLastName)
+      end
+    end
+  end
+end
