+++ lib/api/endpoints/get_by_last_name_descending.rb
@@ -1,0 +1,14 @@
+require 'api/helpers'
+
+module API
+  module Endpoints
+    class GetByLastNameDescending < Grape::API
+      helpers Helpers::SortHelper
+      helpers Helpers::StoreHelper
+
+      get :name do
+        sort(records, :ByLastNameDescending)
+      end
+    end
+  end
+end
