--- lib/api/get_by_gender_and_last_name.rb
@@ -1,13 +1,0 @@
-require 'api/sort_helper'
-require 'api/store_helper'
-
-module API
-  class GetByGenderAndLastName < Grape::API
-    helpers SortHelper
-    helpers StoreHelper
-
-    get :gender do
-      sort(records, :ByGenderAndLastName)
-    end
-  end
-end
