--- lib/api/get_by_last_name_descending.rb
@@ -1,13 +1,0 @@
-require 'api/sort_helper'
-require 'api/store_helper'
-
-module API
-  class GetByLastNameDescending < Grape::API
-    helpers SortHelper
-    helpers StoreHelper
-
-    get :name do
-      sort(records, :ByLastNameDescending)
-    end
-  end
-end
