+++ lib/api/helpers/sort_helper.rb
@@ -1,0 +1,9 @@
+module API
+  module Helpers
+    module SortHelper
+      def sort(records, view)
+        RecordParser::Sorting.const_get(view).new(records).sort
+      end
+    end
+  end
+end
