+++ lib/api/helpers/store_helper.rb
@@ -1,0 +1,9 @@
+module API
+  module Helpers
+    module StoreHelper
+      def records
+        API::Store.records
+      end
+    end
+  end
+end
