--- lib/api/sort_helper.rb
@@ -1,7 +1,0 @@
-module API
-  module SortHelper
-    def sort(records, view)
-      RecordParser::Sorting.const_get(view).new(records).sort
-    end
-  end
-end
