--- lib/api/store_helper.rb
@@ -1,7 +1,0 @@
-module API
-  module StoreHelper
-    def records
-      API::Store.records
-    end
-  end
-end
