+++ spec/api/endpoints/get_by_last_name_descending_spec.rb
@@ -1,0 +1,12 @@
+require 'api_spec_helper'
+require 'support/shared_examples/for_sorting_endpoints'
+
+module API
+  module Endpoints
+    describe GetByLastNameDescending do
+      it_behaves_like 'a sorting endpoint',
+                      'records/name',
+                      RecordParser::Sorting::ByLastNameDescending
+    end
+  end
+end
