--- spec/api/get_by_birth_date_spec.rb
@@ -1,10 +1,0 @@
-require 'api_spec_helper'
-require 'support/shared_examples/for_sorting_endpoints'
-
-module API
-  describe GetByBirthDate do
-    it_behaves_like 'a sorting endpoint',
-                    'records/birthdate',
-                    RecordParser::Sorting::ByBirthDate
-  end
-end
