--- spec/api/get_by_gender_and_last_name_spec.rb
@@ -1,10 +1,0 @@
-require 'api_spec_helper'
-require 'support/shared_examples/for_sorting_endpoints'
-
-module API
-  describe GetByGenderAndLastName do
-    it_behaves_like 'a sorting endpoint',
-                    'records/gender',
-                    RecordParser::Sorting::ByGenderAndLastName
-  end
-end
