--- spec/api/get_by_last_name_descending_spec.rb
@@ -1,10 +1,0 @@
-require 'api_spec_helper'
-require 'support/shared_examples/for_sorting_endpoints'
-
-module API
-  describe GetByLastNameDescending do
-    it_behaves_like 'a sorting endpoint',
-                    'records/name',
-                    RecordParser::Sorting::ByLastNameDescending
-  end
-end
